requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.5.2
Flask-SocketIO==5.3.6
pymongo==4.5.0
python-dotenv==1.0.0
openai==0.28.0
gunicorn==21.2.0
google-generativeai